LayoutTests/imported/w3c/web-platform-tests/css/css-ui/historical/webkit-user-modify-02-expected.txt


FAIL contenteditable="true" doesn't use unsupported -webkit-user-modify property assert_equals: expected "" but got "read-write"
FAIL contenteditable="plaintext-only" doesn't use unsupported -webkit-user-modify property assert_equals: expected "" but got "read-write-plaintext-only"
